requirements.txt
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0
requests>=2.31.0
firebase-admin>=6.2.0
prometheus-client>=0.19.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.13.0
